Added WebIniti & WebSec config, ammended Webconfig
The Websec config is both and might be scrapped at a later stage. I also added a DispatcherServlet in the WebInit config which will act as a Front Controller. The Web Config & Controllers have been amended to reflect this change as they are no longer going to communicate with the outside