This adds the connection to the test database
This adds the settings to the properties file and the Column annotations for all the correct fields. Spring does not currently support usage without the annotation because naming conventions for Postgres were not adhered to in the underlying database. This is also the source for all the escaped quotation marks