Add the functionality required by the frontend.
This starts implementing Streams to make the database access not as much of a hinderance as it was before. It can now start processing data as soon as it starts finding the first, making this process quite a bit faster. Still takes about 4 times as long as the lightweight version with overloaded service methods (will be implemented). Comparison for a 15k CSV with about 100 Datastreams: 13 vs 3 Seconds.